Google Play policy update compliance.
Apps for API29 that allow unrestricted package enumeration will be banned for upload to Play in Nov 2021.
QUERY_ALL_PACKAGES permission (that allows enumeration on API30) usage will be restricted in March 2022, so in order to continue to be able to access installed skins we have to implement this dummy receiver.